Tabluated entries & removed some content
Removed some duplicates, added some url links

Removed some coingecko html code present at the end of some descriptions that showed the Ethereum logo on these entries.

Tabulated (markdown editing) the content of sections that were missing. Did not tabulate 'Other' as these will probably be added to other sections in future. 

Took the opportunity to move the entire list that is the 'Other' section at the end, which ends before the 'Events' section.  Seemed the most logical arrangement for the content